+++ river_route/routers/__init__.py
@@ -1,3 +1,4 @@
-from .Router import Router
+from .Config import Configs
+from .Muskingum import Muskingum
 from .RapidMuskingum import RapidMuskingum
 from .UnitMuskingum import UnitMuskingum